+++ CHANGELOG.md
@@ -37,9 +37,9 @@ Adding a new version? You'll need three changes:
 
 ### Added
 
-- Added `listenerOptions` in `GatewayConfiguration`'s `spec` to allow overriding
+- Added `listenersOptions` in `GatewayConfiguration`'s `spec` to allow overriding
   the default configuration related to a certain listener in the `Gateway`.
-  Each item of `listenerOptions` matches a listener in the `Gateway` which the
+  Each item of `listenersOptions` matches a listener in the `Gateway` which the
   `GatewayConfiguration` configures by its `name`. The listeners without a
   matching item in `GatewayConfiguration` remains the default configuration.
   Currently `nodePort` can be configured to set the exposed nodeport of the